R0.3.0 changes

+++ fix-jlcpcb-files.sh
@@ -8,12 +8,15 @@ DESTDIR="assembly"
 BOMFILE="${PROJ}_bom_jlc.csv"
 CPLFILE="${PROJ}_cpl_jlc.csv"
 
-# fix the BOM and CPL files
-jlc-kicad-tools -n $PROJ -d cpl_rotations_db.csv .
+# if we're ordering PCBs only, ignore the BOM and CPL parts
+if [ -f $BOMFILE ] ; then
+    # fix the BOM and CPL files
+    jlc-kicad-tools -n $PROJ -d cpl_rotations_db.csv .
 
-# move the BOM and CPL files
-mkdir -p $DESTDIR
-mv $BOMFILE $CPLFILE $DESTDIR
+    # move the BOM and CPL files
+    mkdir -p $DESTDIR
+    mv $BOMFILE $CPLFILE $DESTDIR
+fi
 
 # zip the DESTDIR directory
 zip -FSr $DESTDIR.zip $DESTDIR/
